requirements/base.txt
Django==1.5.5
django-cms==2.4.3
django-blog-zinnia==0.12.3
django-autoslug
django-social-auth
pytils
sorl-thumbnail
South
Pillow
